refactor: :art: Moved hooks folder
Moved hooks folder to src folder.

hooks
