+++ src/components/index.ts
@@ -3,5 +3,3 @@ export { default as Thumbnail } from './Thumbnail/Thumbnail'
 export { default as Detail } from './Detail/Detail'
 export { default as Layout } from './Layout/Layout'
 export { default as LoadingSpinner } from './LoadingSpinner/LoadingSpinner'
-export { default as useLocalStorage } from './hooks/useLocalStorage'
-export { default as useWindowSize } from './hooks/useWindowSize'
