+++ src/hooks/index.ts
@@ -1,0 +1,2 @@
+export { default as useLocalStorage } from './useLocalStorage'
+export { default as useWindowSize } from './useWindowSize'
