+++ src/pages/MovieDetail.tsx
@@ -3,7 +3,8 @@ import { useParams, useLocation } from 'react-router-dom'
 import { useQuery } from 'react-query'
 import { Alert, AlertTitle } from '@material-ui/lab'
 
-import { Detail, LoadingSpinner, useLocalStorage } from '../components'
+import { Detail, LoadingSpinner } from '../components'
+import { useLocalStorage } from '../hooks'
 import { setMovies, hasMovie } from '../utils/helpers'
 import { IMovie, IMovieDetail, IResponseError } from '../utils/types'
 import { config } from '../utils'
